+++ crates/rustc_codegen_spirv/src/builder_spirv.rs
@@ -13,6 +13,7 @@ use rustc_arena::DroplessArena;
 use rustc_data_structures::fx::{FxHashMap, FxHashSet};
 use rustc_data_structures::sync::Lrc;
 use rustc_middle::bug;
+use rustc_middle::mir::interpret::ConstAllocation;
 use rustc_middle::ty::TyCtxt;
 use rustc_span::source_map::SourceMap;
 use rustc_span::symbol::Symbol;
@@ -206,8 +207,8 @@ impl SpirvValueExt for Word {
     }
 }
 
-#[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, Hash)]
-pub enum SpirvConst<'tcx> {
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+pub enum SpirvConst<'a, 'tcx> {
     U32(u32),
     U64(u64),
     /// f32 isn't hash, so store bits
@@ -225,16 +226,23 @@ pub enum SpirvConst<'tcx> {
     // different functions, but of the same type, don't overlap their zombies.
     ZombieUndefForFnAddr,
 
-    Composite(&'tcx [Word]),
+    Composite(&'a [Word]),
 
     /// Pointer to constant data, i.e. `&pointee`, represented as an `OpVariable`
     /// in the `Private` storage class, and with `pointee` as its initializer.
     PtrTo {
         pointee: Word,
     },
+
+    /// Symbolic result for the `const_data_from_alloc` method, to allow deferring
+    /// the actual value generation until after a pointer to this value is cast
+    /// to its final type (e.g. that will be loaded as).
+    //
+    // FIXME(eddyb) replace this with `qptr` handling of constant data.
+    ConstDataFromAlloc(ConstAllocation<'tcx>),
 }
 
-impl SpirvConst<'_> {
+impl<'tcx> SpirvConst<'_, 'tcx> {
     /// Replace `&[T]` fields with `&'tcx [T]` ones produced by calling
     /// `tcx.arena.dropless.alloc_slice(...)` - this is done late for two reasons:
     /// 1. it avoids allocating in the arena when the cache would be hit anyway,
@@ -242,7 +250,7 @@ impl SpirvConst<'_> {
     ///    (ideally these would also be interned, but that's even more refactors)
     /// 2. an empty slice is disallowed (as it's usually handled as a special
     ///    case elsewhere, e.g. `rustc`'s `ty::List` - sadly we can't use that)
-    fn tcx_arena_alloc_slices<'tcx>(self, cx: &CodegenCx<'tcx>) -> SpirvConst<'tcx> {
+    fn tcx_arena_alloc_slices(self, cx: &CodegenCx<'tcx>) -> SpirvConst<'tcx, 'tcx> {
         fn arena_alloc_slice<'tcx, T: Copy>(cx: &CodegenCx<'tcx>, xs: &[T]) -> &'tcx [T] {
             if xs.is_empty() {
                 &[]
@@ -264,6 +272,8 @@ impl SpirvConst<'_> {
             SpirvConst::PtrTo { pointee } => SpirvConst::PtrTo { pointee },
 
             SpirvConst::Composite(fields) => SpirvConst::Composite(arena_alloc_slice(cx, fields)),
+
+            SpirvConst::ConstDataFromAlloc(alloc) => SpirvConst::ConstDataFromAlloc(alloc),
         }
     }
 }
@@ -282,6 +292,12 @@ enum LeafIllegalConst {
     /// as its operands, and `OpVariable`s are never considered constant.
     // FIXME(eddyb) figure out if this is an accidental omission in SPIR-V.
     CompositeContainsPtrTo,
+
+    /// `ConstDataFromAlloc` constant, which cannot currently be materialized
+    /// to SPIR-V (and requires to be wrapped in `PtrTo` and bitcast, first).
+    //
+    // FIXME(eddyb) replace this with `qptr` handling of constant data.
+    UntypedConstDataFromAlloc,
 }
 
 impl LeafIllegalConst {
@@ -289,6 +305,10 @@ impl LeafIllegalConst {
         match *self {
             Self::CompositeContainsPtrTo => {
                 "constant arrays/structs cannot contain pointers to other constants"
+            }
+            Self::UntypedConstDataFromAlloc => {
+                "`const_data_from_alloc` result wasn't passed through `static_addr_of`, \
+                 then `const_bitcast` (which would've given it a type)"
             }
         }
     }
@@ -391,8 +411,8 @@ pub struct BuilderSpirv<'tcx> {
     // (e.g. `OpConstant...`) instruction.
     // NOTE(eddyb) both maps have `WithConstLegality` around their keys, which
     // allows getting that legality information without additional lookups.
-    const_to_id: RefCell<FxHashMap<WithType<SpirvConst<'tcx>>, WithConstLegality<Word>>>,
-    id_to_const: RefCell<FxHashMap<Word, WithConstLegality<SpirvConst<'tcx>>>>,
+    const_to_id: RefCell<FxHashMap<WithType<SpirvConst<'tcx, 'tcx>>, WithConstLegality<Word>>>,
+    id_to_const: RefCell<FxHashMap<Word, WithConstLegality<SpirvConst<'tcx, 'tcx>>>>,
 
     debug_file_cache: RefCell<FxHashMap<DebugFileKey, DebugFileSpirv<'tcx>>>,
 
@@ -535,7 +555,7 @@ impl<'tcx> BuilderSpirv<'tcx> {
     pub(crate) fn def_constant_cx(
         &self,
         ty: Word,
-        val: SpirvConst<'_>,
+        val: SpirvConst<'_, 'tcx>,
         cx: &CodegenCx<'tcx>,
     ) -> SpirvValue {
         let val_with_type = WithType { ty, val };
@@ -564,7 +584,9 @@ impl<'tcx> BuilderSpirv<'tcx> {
             }
 
             SpirvConst::Null => builder.constant_null(ty),
-            SpirvConst::Undef | SpirvConst::ZombieUndefForFnAddr => builder.undef(ty, None),
+            SpirvConst::Undef
+            | SpirvConst::ZombieUndefForFnAddr
+            | SpirvConst::ConstDataFromAlloc(_) => builder.undef(ty, None),
 
             SpirvConst::Composite(v) => builder.constant_composite(ty, v.iter().copied()),
 
@@ -635,6 +657,10 @@ impl<'tcx> BuilderSpirv<'tcx> {
                     Err(IllegalConst::Indirect(cause))
                 }
             },
+
+            SpirvConst::ConstDataFromAlloc(_) => Err(IllegalConst::Shallow(
+                LeafIllegalConst::UntypedConstDataFromAlloc,
+            )),
         };
         let val = val.tcx_arena_alloc_slices(cx);
         assert_matches!(
@@ -658,11 +684,11 @@ impl<'tcx> BuilderSpirv<'tcx> {
         SpirvValue { kind, ty }
     }
 
-    pub fn lookup_const_by_id(&self, id: Word) -> Option<SpirvConst<'tcx>> {
+    pub fn lookup_const_by_id(&self, id: Word) -> Option<SpirvConst<'tcx, 'tcx>> {
         Some(self.id_to_const.borrow().get(&id)?.val)
     }
 
-    pub fn lookup_const(&self, def: SpirvValue) -> Option<SpirvConst<'tcx>> {
+    pub fn lookup_const(&self, def: SpirvValue) -> Option<SpirvConst<'tcx, 'tcx>> {
         match def.kind {
             SpirvValueKind::Def(id) | SpirvValueKind::IllegalConst(id) => {
                 self.lookup_const_by_id(id)